Ensure integer-only equations, add markdown rendering, and hide equations during exams (#28)
* Initial plan

* Implement integer-only equations, number formatting, markdown rendering, and visibility improvements


* Add comprehensive tests and fix linter issues


* fix

---------

+++ models.py
@@ -29,7 +29,7 @@ class Question(BaseModel):
     """
     equation: str
     question_text: str
-    answer: float
+    answer: int
     difficulty: str
     category: Optional[str] = None
